+++ src/Exception/Route/InvalidRoutableException.php
@@ -1,9 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Torr\Rad\Exception\Route;
-
-use Torr\Rad\Exception\UnexpectedTypeException;
-
-final class InvalidRoutableException extends UnexpectedTypeException
-{
-}