+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * (internal) Use HtmlBuilder for `DataContainer`.
 * (bc) Replace `IdTrait` + `TimestampsTrait` to `EntityFieldsTrait` and `ModifiableEntityFieldsTrait`.
 * (bc) Moved `EntityInterface` to new namespace and add time created getter.
+* (bc) Remove `Routable`.
 
 
 2.6.1